Extract helpers from config.rb
